Move the batched layer driver into driver.rs
mod.rs was 690 lines: the batched layer schedule plus the tests for it.
Every other file in the module already held one concern and its own tests.
This was the last piece that had not been separated.

BatchProveOutput, combine_claims, reduce_layer with its four steps, and
batch_prove_unequal_depths move to driver.rs, with the tests that cover
them. mod.rs is left at 38 lines: the module doc and the re-exports.

Three tests went elsewhere, because mod.rs cannot keep a test module and
still be documentation and re-exports. They were placed by what they
cover, not by where they came from: the two prove-verify round trips
exercise FracAddCircuit::prove, so they joined circuit.rs; the pad/unpad
inverse proptest exercises unpad_leaf_claim, so it opened padding.rs's
first test module.

mod driver is private with a re-export, so
fracaddcheck::batch_prove_unequal_depths and fracaddcheck::BatchProveOutput
keep their paths. logup_star/prove.rs and the bench are unmodified.

The module doc also records the one layout decision whose obvious
alternative is slower: each instance of a batched layer keeps its own
column store, and folding them into one costs 1.5-2x. That is the front
door for this module, so it is where a reader will look before making the
change that looks natural.

The transcript is byte-identical: dumping transcript.finalize() plus the
returned eval_point and fractions at depths [2,4,5], [4,4,4] and [0,3]
gives sha256 b939c8e2...06201 before and after. Independently, the moved
non-test code diffs clean against the old mod.rs apart from two doc-link
paths, and the moved test bodies diff clean too.

Test count is 94 before and after, in both rayon modes.
